+++ like/src/timer/timer.ts
@@ -7,7 +7,6 @@ export class Timer {
   private frameCount = 0;
   private fps = 0;
   private fpsAccumulator = 0;
-  private sleepUntil: number | null = null;
 
   constructor(props: EngineProps<never>) {
     props.canvas.addEventListener("like:update", this.update.bind(this), { signal: props.abort })
@@ -41,29 +40,5 @@ export class Timer {
   /** Get the ingame time. */
   getTime(): number {
     return this.totalTime;
-  }
-
-  /**
-   * Whether or not the game is (supposed to be) frozen.
-   * The only callback while sleeping is `draw`, and
-   * calling this outside of `draw` will always return
-   * false -- except if you have a custom runtime.
-   */
-  isSleeping(): boolean {
-    if (this.sleepUntil === null) return false;
-    const currentTime = performance.now();
-    if (currentTime < this.sleepUntil) {
-      return true;
-    }
-    this.sleepUntil = null;
-    return false;
-  }
-
-  /**
-   * Freeze the whole game for a time. Audio will keep playing,
-   * but update functions won't be called and events won't fire.
-   */
-  sleep(duration: number): void {
-    this.sleepUntil = performance.now() + (duration * 1000);
   }
 }